Cleanup of the XRender fade plugin.
Changes:
> Windows can now return their dimensions as an XRectangle for
  storage/reference convenience.
> Plugins can now access the Display pointer with an appropriate
  function.
> Rewrote a big part of the XRender fade plugin logic. The plugin itself
  should be much cleaner overall.
> Other fixes.

+++ util/fbcompose/XRenderWindow.cc
@@ -113,7 +113,7 @@ void XRenderWindow::updateMaskPicture() throw() {
     XRenderColor color = { 0, 0, 0, 0 };
     XRenderFillRectangle(display(), PictOpSrc, direct_maskPicture(), &color, 0, 0, realWidth(), realHeight());
 
-    color.alpha = (unsigned int)((alpha() * 0xffff) / 255.0);
+    color.alpha = (unsigned long)((alpha() * 0xffff) / 255.0);
     XRenderFillRectangles(display(), PictOpSrc, direct_maskPicture(), &color, clipShapeRects(), clipShapeRectCount());
 }
